Heavy Enemy added
New type of enemy added: HeavyEnemy. If the player is inside the room of the enemy (something that is calculated by two vectors that represents the coordinates bounds of the room), the enemy will seek the player and its velocity will increase each second the player is at the room. The enemy type selection was also improved by using an enum and a switch, instead of multiple bools and ifs.